Migrate access durations to configuration
Bring it in line with other configurations
