+++ backend/app/configuration/access_durations/schema_response.py
@@ -1,7 +1,7 @@
 from uuid import UUID
 
 from app.abstract_data_product.type import AbstractDataProductType
-from app.access_durations.enums import AccessDurationType
+from app.configuration.access_durations.enums import AccessDurationType
 from app.shared.schema import ORMModel
