+++ backend/tests/app/data_products/output_ports/input_ports/test_router.py
@@ -9,9 +9,9 @@ from fastapi.testclient import TestClient
 from httpx import Response
 
 from app.abstract_data_product.type import AbstractDataProductType
-from app.access_durations.enums import AccessDurationType
 from app.authorization.role_assignments.enums import DecisionStatus
 from app.authorization.roles.schema import Scope
+from app.configuration.access_durations.enums import AccessDurationType
 from app.core.authz import Action
 from app.data_products.output_ports.enums import OutputPortAccessType
 from app.settings import settings
