+++ docs/docs/release-notes.md
@@ -6,6 +6,10 @@ sidebar_position: 200
 
 ## Unreleased
 
+### breaking changes
+
+- **[API]**: Access durations configuration switched URL path from /v2/access_durations to /v2/configuration/access_durations. Please update your API calls accordingly.
+
 ## 0.7.0
 
 ### breaking changes
