+++ frontend/src/components/abstract-data-products/input-port-tab/components/input-port-table/input-port-action-button.component.tsx
@@ -3,7 +3,7 @@ import { Button, Flex, Popconfirm } from 'antd';
 import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DEFAULT_EXPIRING_SOON_THRESHOLD_DAYS, isExpiringSoon } from '@/components/input-port/access-status.tsx';
-import { useGetExpiringSoonThresholdQuery } from '@/store/api/services/generated/accessDurationsApi.ts';
+import { useGetExpiringSoonThresholdQuery } from '@/store/api/services/generated/configurationAccessDurationsApi.ts';
 import { InputPortStatus, type OutputPort, RenewalStatus } from '@/store/api/services/generated/dataProductsApi.ts';
 import { dispatchMessage } from '@/store/features/feedback/utils/dispatch-feedback.ts';
