+++ frontend/src/components/datasets/dataset-form/dataset-form.component.tsx
@@ -25,14 +25,14 @@ import { useDebouncedCallback } from 'use-debounce';
 import { ResourceNameFormItem } from '@/components/resource-name/resource-name-form-item.tsx';
 import { FORM_GRID_WRAPPER_COLS, MAX_DESCRIPTION_INPUT_LENGTH } from '@/constants/form.constants.ts';
 import { TabKeys } from '@/pages/data-product/components/data-product-tabs/data-product-tabkeys';
+import { useCheckAccessQuery } from '@/store/api/services/generated/authorizationApi.ts';
 import {
     AbstractDataProductType,
     type AccessDuration,
     AccessDurationType,
     useGetAllAccessDurationsQuery,
     useIsTimeBoundAccessEnabledQuery,
-} from '@/store/api/services/generated/accessDurationsApi.ts';
-import { useCheckAccessQuery } from '@/store/api/services/generated/authorizationApi.ts';
+} from '@/store/api/services/generated/configurationAccessDurationsApi.ts';
 import { useGetDataProductsLifecyclesQuery } from '@/store/api/services/generated/configurationDataProductLifecyclesApi.ts';
 import { useGetTagsQuery } from '@/store/api/services/generated/configurationTagsApi.ts';
 import { OutputPortAccessType, useGetDataProductQuery } from '@/store/api/services/generated/dataProductsApi.ts';
