+++ frontend/src/store/api/services/apiTags.ts
@@ -1,6 +1,6 @@
 import { api } from '@/store/api/services/generated/completeServiceApi.ts';
 import { STATIC_TAG_ID, TagTypes } from '@/store/api/services/tag-types.ts';
-import { accessDurationsTags } from '@/store/api/services/tags/accessDurationsTags.ts';
+import { configurationAccessDurationsTags } from '@/store/api/services/tags/configurationAccessDurationsTags.ts';
 import { dataProductsOutputPortsInputPortsTags } from '@/store/api/services/tags/dataProductsOutputPortsInputPortsTags.ts';
 import { dataProductOutputPortTags } from '@/store/api/services/tags/dataProductsOutputPortsTags.ts';
 import { dataProductTechnicalAssetsTags } from '@/store/api/services/tags/dataProductsTechicalAssetsTags.ts';
@@ -299,6 +299,6 @@ api.enhanceEndpoints({
         ...usersTags,
         ...dataProductsOutputPortsInputPortsTags,
         ...explorationTags,
-        ...accessDurationsTags,
+        ...configurationAccessDurationsTags,
     },
 });
