+++ sdk/sdk/api_client/api/configuration_access_durations/get_all_access_durations.py
@@ -5,7 +5,7 @@ import httpx
 
 from ... import errors
 from ...client import AuthenticatedClient, Client
-from ...models.time_bound_access_enabled_response import TimeBoundAccessEnabledResponse
+from ...models.access_duration import AccessDuration
 from ...types import Response
 
 
@@ -13,7 +13,7 @@ def _get_kwargs() -> dict[str, Any]:
 
     _kwargs: dict[str, Any] = {
         "method": "get",
-        "url": "/api/v2/access_durations/enabled",
+        "url": "/api/v2/configuration/access_durations",
     }
 
     return _kwargs
@@ -21,9 +21,14 @@ def _get_kwargs() -> dict[str, Any]:
 
 def _parse_response(
     *, client: AuthenticatedClient | Client, response: httpx.Response
-) -> TimeBoundAccessEnabledResponse | None:
+) -> list[AccessDuration] | None:
     if response.status_code == 200:
-        response_200 = TimeBoundAccessEnabledResponse.from_dict(response.json())
+        response_200 = []
+        _response_200 = response.json()
+        for response_200_item_data in _response_200:
+            response_200_item = AccessDuration.from_dict(response_200_item_data)
+
+            response_200.append(response_200_item)
 
         return response_200
 
@@ -35,7 +40,7 @@ def _parse_response(
 
 def _build_response(
     *, client: AuthenticatedClient | Client, response: httpx.Response
-) -> Response[TimeBoundAccessEnabledResponse]:
+) -> Response[list[AccessDuration]]:
     return Response(
         status_code=HTTPStatus(response.status_code),
         content=response.content,
@@ -47,15 +52,15 @@ def _build_response(
 def sync_detailed(
     *,
     client: AuthenticatedClient | Client,
-) -> Response[TimeBoundAccessEnabledResponse]:
-    """Is Time Bound Access Enabled
+) -> Response[list[AccessDuration]]:
+    """Get All Access Durations
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[TimeBoundAccessEnabledResponse]
+        Response[list[AccessDuration]]
     """
 
     kwargs = _get_kwargs()
@@ -70,15 +75,15 @@ def sync_detailed(
 def sync(
     *,
     client: AuthenticatedClient | Client,
-) -> TimeBoundAccessEnabledResponse | None:
-    """Is Time Bound Access Enabled
+) -> list[AccessDuration] | None:
+    """Get All Access Durations
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        TimeBoundAccessEnabledResponse
+        list[AccessDuration]
     """
 
     return sync_detailed(
@@ -89,15 +94,15 @@ def sync(
 async def asyncio_detailed(
     *,
     client: AuthenticatedClient | Client,
-) -> Response[TimeBoundAccessEnabledResponse]:
-    """Is Time Bound Access Enabled
+) -> Response[list[AccessDuration]]:
+    """Get All Access Durations
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[TimeBoundAccessEnabledResponse]
+        Response[list[AccessDuration]]
     """
 
     kwargs = _get_kwargs()
@@ -110,15 +115,15 @@ async def asyncio_detailed(
 async def asyncio(
     *,
     client: AuthenticatedClient | Client,
-) -> TimeBoundAccessEnabledResponse | None:
-    """Is Time Bound Access Enabled
+) -> list[AccessDuration] | None:
+    """Get All Access Durations
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        TimeBoundAccessEnabledResponse
+        list[AccessDuration]
     """
 
     return (
